+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.7.0
 
 * Curl and Kerl algorithm updated to use raw number array for speed
+* Kerl rewritten to match java implementation
+* Sha3/Keccak implementation added
 * Added code coverage to testing